Add files via upload
In this project, we have successfully implemented ARP poisoning, DOS, and DNS poisoning attacks. We have also recorded the procedure and explained it step-by-step in great detail. Even though these attacks have a limitation of only working on LAN, these are very common and easily performed and virtually undetectable by a novice. As a result, this is a threat that has to be addressed. So, we have come up with an app “anti-Dns” which is capable of detecting, preventing and recovering from an ARP poisoning attack on the victim’s Linux system. This is written in python and mainly focuses on analyzing the ARP cache of the system to detect attacks. Also there are functions which have the capability of performing continuous scans after a certain regular interval of time.